[01] Move transformations into separate module
